+++ pyzoo/zoo/zouwu/autots/forecast.py
@@ -16,7 +16,7 @@
 
 from zoo.automl.regression.time_sequence_predictor import TimeSequencePredictor
 from zoo.zouwu.config.recipe import *
-from zoo.automl.pipeline.time_sequence import load_ts_pipeline
+from zoo.zouwu.pipeline.time_sequence import load_ts_pipeline
 
 
 class AutoTSTrainer:
